Removing code duplication
