+++ .github/workflows/tensorflow.yml
@@ -26,6 +26,7 @@ jobs:
         run: |
           python -m pip install --upgrade pip
           pip install GitPython
+          pip install . --no-deps
       - name: Get changed backends
         id: backend_check
         run: |
