+++ .github/workflows/xgboost.yml
@@ -23,6 +23,7 @@ jobs:
         run: |
           python -m pip install --upgrade pip
           pip install GitPython
+          pip install . --no-deps
       - name: Get changed backends
         id: backend_check
         run: |
